+++ godot/addons/cyclops_level_builder/actions/action_export_as_gltf.gd
@@ -25,12 +25,12 @@
 class_name ActionExportAsGltf
 extends CyclopsAction
 
+var wizard:ExporterGltfWizard = preload("res://addons/cyclops_level_builder/exporter/exporter_gltf_wizard.tscn").instantiate()
 
 func _init(plugin:CyclopsLevelBuilder, name:String = "", accellerator:Key = KEY_NONE):
 	super._init(plugin, "Export As Gltf...")
 
 func _execute():
-	var wizard:ExporterGltfWizard = preload("res://addons/cyclops_level_builder/exporter/exporter_gltf_wizard.tscn").instantiate()
 	var base_control:Node = plugin.get_editor_interface().get_base_control()
 	base_control.add_child(wizard)