+++ godot/addons/cyclops_level_builder/commands/cmd_transform_blocks.gd
@@ -60,14 +60,14 @@ func do_it():
 		var new_w_xform:Transform3D = transform * w_init_xform
 		block.global_transform = new_w_xform
 		
-		#if !lock_uvs:
-			#var vol:ConvexVolume = ConvexVolume.new()
-			#vol.init_from_convex_block_data(tracked.data)
-			#
-			#var uv_xform:Transform3D = transform.affine_inverse()
-			#vol.transform_uvs(uv_xform)
-			#
-			#block.block_data = vol.to_convex_block_data()
+		if !lock_uvs:
+			var vol:ConvexVolume = ConvexVolume.new()
+			vol.init_from_convex_block_data(tracked.data)
+			
+			var uv_xform:Transform3D = transform.affine_inverse()
+			vol.transform_uvs(uv_xform)
+			
+			block.block_data = vol.to_convex_block_data()
 
 func undo_it():
 	for tracked in tracked_blocks: