+++ godot/addons/cyclops_level_builder/menu/editor_toolbar.gd
@@ -54,6 +54,7 @@ func _ready():
 	%Menu.add_separator()
 	%Menu.add_action_item(ActionConvertToMesh.new(editor_plugin))
 	%Menu.add_action_item(ActionExportAsGltf.new(editor_plugin))
+	%Menu.add_action_item(ActionExportAsGodotScene.new(editor_plugin))
 	%Menu.add_separator()
 	%Menu.add_action_item(ActionRotateX90Ccw.new(editor_plugin))
 	%Menu.add_action_item(ActionRotateX90Cw.new(editor_plugin))