+++ doc/change_log.md
@@ -6,3 +6,8 @@
 * Shape creation tools now let you specify collision settings.
 * Right click cancel now does not undo rotation and move after operation has been completed.
 * Now flushing snapping changes whenever a setting is changed.
+* Renaming Settings to CyclopsSettings
+* Now generating tangent vectors for block meshes.
+* Restoring ability for UVs to remain in place when moving blocks.
+* Adding option to clear UV transform in material brush.
+* Adding tool to export current scene as Godot scene with only native Godot objects.